refactor(exporter-shopware-6) Rename Category mapper
