+++ module/exporter-shopware-6/src/Infrastructure/Builder/CategoryBuilder.php
@@ -6,30 +6,29 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Category;
+namespace Ergonode\ExporterShopware6\Infrastructure\Builder;
 
 use Ergonode\Category\Domain\Entity\AbstractCategory;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Domain\Repository\CategoryRepositoryInterface;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CategoryMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CategoryMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Category;
 use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 
-class Shopware6CategoryParentMapper implements Shopware6CategoryMapperInterface
+class CategoryBuilder
 {
-    private CategoryRepositoryInterface $repository;
+    /**
+     * @var CategoryMapperInterface[]
+     */
+    private array $collection;
 
-    public function __construct(CategoryRepositoryInterface $repository)
+    public function __construct(CategoryMapperInterface ...$collection)
     {
-        $this->repository = $repository;
+        $this->collection = $collection;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function map(
+    public function build(
         Shopware6Channel $channel,
         Export $export,
         Shopware6Category $shopware6Category,
@@ -37,9 +36,15 @@ class Shopware6CategoryParentMapper implements Shopware6CategoryMapperInterface
         ?CategoryId $parentCategoryId = null,
         ?Language $language = null
     ): Shopware6Category {
-        if ($parentCategoryId) {
-            $shopwareParentId = $this->repository->load($channel->getId(), $parentCategoryId);
-            $shopware6Category->setParentId($shopwareParentId);
+        foreach ($this->collection as $mapper) {
+            $shopware6Category = $mapper->map(
+                $channel,
+                $export,
+                $shopware6Category,
+                $category,
+                $parentCategoryId,
+                $language
+            );
         }
 
         return $shopware6Category;
