+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Category/CategoryNameMapper.php
@@ -6,17 +6,21 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\ExporterShopware6\Infrastructure\Mapper;
+namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Category;
 
 use Ergonode\Category\Domain\Entity\AbstractCategory;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CategoryMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Category;
 use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 
-interface Shopware6CategoryMapperInterface
+class CategoryNameMapper implements CategoryMapperInterface
 {
+    /**
+     * {@inheritDoc}
+     */
     public function map(
         Shopware6Channel $channel,
         Export $export,
@@ -24,5 +28,16 @@ interface Shopware6CategoryMapperInterface
         AbstractCategory $category,
         ?CategoryId $parentCategoryId = null,
         ?Language $language = null
-    ): Shopware6Category;
+    ): Shopware6Category {
+        $name = $category->getName()->get($language ?: $channel->getDefaultLanguage());
+        if ($name) {
+            $shopware6Category->setName($name);
+        }
+
+        if (null === $language && null === $name) {
+            $shopware6Category->setName($category->getCode()->getValue());
+        }
+
+        return $shopware6Category;
+    }
 }
