+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Category/CategoryParentMapper.php
@@ -6,29 +6,30 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\ExporterShopware6\Infrastructure\Builder;
+namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Category;
 
 use Ergonode\Category\Domain\Entity\AbstractCategory;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Channel\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6CategoryMapperInterface;
+use Ergonode\ExporterShopware6\Domain\Repository\CategoryRepositoryInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\CategoryMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Category;
 use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 
-class Shopware6CategoryBuilder
+class CategoryParentMapper implements CategoryMapperInterface
 {
-    /**
-     * @var Shopware6CategoryMapperInterface[]
-     */
-    private array $collection;
+    private CategoryRepositoryInterface $repository;
 
-    public function __construct(Shopware6CategoryMapperInterface ...$collection)
+    public function __construct(CategoryRepositoryInterface $repository)
     {
-        $this->collection = $collection;
+        $this->repository = $repository;
     }
 
-    public function build(
+    /**
+     * {@inheritDoc}
+     */
+    public function map(
         Shopware6Channel $channel,
         Export $export,
         Shopware6Category $shopware6Category,
@@ -36,15 +37,9 @@ class Shopware6CategoryBuilder
         ?CategoryId $parentCategoryId = null,
         ?Language $language = null
     ): Shopware6Category {
-        foreach ($this->collection as $mapper) {
-            $shopware6Category = $mapper->map(
-                $channel,
-                $export,
-                $shopware6Category,
-                $category,
-                $parentCategoryId,
-                $language
-            );
+        if ($parentCategoryId) {
+            $shopwareParentId = $this->repository->load($channel->getId(), $parentCategoryId);
+            $shopware6Category->setParentId($shopwareParentId);
         }
 
         return $shopware6Category;
